requirements.txt
absl-py
tensorflow
pandas
numpy
scikit-learn
